% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{farsight-package}
\alias{farsight}
\alias{farsight-package}
\title{farsight - A package to retrieve intelligence from Farsight DNSDB API}
\description{
farsight - A package to retrieve intelligence from Farsight DNSDB API
}
\author{
Jay Jacobs <jay.jacobs@verizon.com>, Bob Rudis <bob.rudis@verizon.com>
}

