% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getresult}
\alias{getresult}
\title{Internal function: given a request for data, returns a data frame of the results}
\usage{
getresult(req)
}
\arguments{
\item{req}{the httr GET request}
}
\description{
Internal function: given a request for data, returns a data frame of the results
}

