% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Parse.INI}
\alias{Parse.INI}
\title{Internal function to load an INI file}
\usage{
Parse.INI(INI.filename)
}
\arguments{
\item{INI.filename}{the filename of the ini file.}
}
\description{
via http://r.789695.n4.nabble.com/Read-Windows-like-INI-files-into-R-data-structure-td827353.html
}

