% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rdata}
\alias{rdata}
\title{Perform an "inverse" lookup based on RData records.}
\usage{
rdata(value, type = "ip", rrtype = NULL, limit = NULL, pat = NULL)
}
\arguments{
\item{value}{depending on the type, this is the object to query for.}

\item{type}{either "ip" or "name"}

\item{rrtype}{optional filter of DNS RRtype field}

\item{limit}{the maximum number of values to return}

\item{pat}{the DNSDB API key (if supplied manually) to use, otherwise this
will try to pull what was loaded from \link{readConfig}, and if that doesn't work
and the session is interactive, it will prompt for the API key.}
}
\description{
Looks up the Rdata index given a type and name and it can optionally filter based on the
RRType or the bailiwick scope.
}
\details{
This will attempt to convert the date/time fields to POSIXct objects and if those are
found, it will attempt to calculate the difference (in days) between the last seen and
first seen record.  Keep an eye out because some of the return values in the columns may
be lists and not easy vectors, specifically the \code{rdata} column.
}
\examples{
# pull names that resolved to an IP
bp <- rdata("113.10.174.118")

# look at 10 most recent rrnames at a given name server
bp <- rdata("ns5.value-domain.com", type="name", limit=10)
}
\seealso{
\url{https://api.dnsdb.info/}
}

