% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readConfig}
\alias{readConfig}
\title{Load the API key for DNSDB API}
\usage{
readConfig(configFilePath)
}
\arguments{
\item{configFilePath}{path to the configuration file}
}
\description{
This reads in a an "analysis.cfg" Python configparser file with API keys and
extracts the key for dnsdp_api into an environment variable for use in the
API.
}
\details{
The file could have one line that loos like this:
dnsdb_api=0dc45ef13ebbe51188e7a7778c019e7c4fe41db9c43ea06866297c88b42bad35
}
\examples{
\dontrun{
readConfig("~/.analysis.cfg")
}
}

