% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rrset}
\alias{rrset}
\title{Perform a "forward" lookup based on the owner name of an RRset.}
\usage{
rrset(name, rrtype = NULL, bailiwick = NULL, limit = NULL, pat = NULL)
}
\arguments{
\item{name}{the domain being searched, may contain wildcard "*" symbol}

\item{rrtype}{optional filter of DNS RRtype field}

\item{bailiwick}{optional filter for bailiwick}

\item{limit}{the maximum number of values to return}

\item{pat}{the DNSDB API key (if supplied manually) to use, otherwise this
will try to pull what was loaded from \link{readConfig}, and if that doesn't work
and the session is interactive, it will prompt for the API key.}
}
\description{
Look up the RRset data for a given name, can optionally filter based on the
RRType or the bailiwick scope.
}
\details{
This will attempt to convert the date/time fields to POSIXct objects and if those are
found, it will attempt to calculate the difference (in days) between the last seen and
first seen record.  Keep an eye out because some of the return values in the columns may
be lists and not easy vectors, specifically the \code{rdata} column.
}
\examples{
# get everything for a simple domain
bp <- rrset("beechplane.com")

# just get the SOA record for it
bp <- rrset("beechplane.com", rrtype="SOA")

# get just the name servers from the top level com domain
bp <- rrset("beechplane.com", rrtype="NS", bailiwick="com.")
}
\seealso{
\url{https://api.dnsdb.info/}
}

